+++ src/main/java/com/home/SteticApi/controller/OrderController.java
@@ -5,6 +5,7 @@ import com.home.SteticApi.domain.ErrorResponse;
 import com.home.SteticApi.domain.Order;
 import com.home.SteticApi.domain.Product;
 import com.home.SteticApi.dto.OrderInDto;
+import com.home.SteticApi.dto.OrderOutDto;
 import com.home.SteticApi.exception.ClientException.ClientNotFoundException;
 import com.home.SteticApi.exception.OrderException.OrderNotFoundException;
 import com.home.SteticApi.exception.ProductException.ProductNotFoundException;
@@ -25,7 +26,8 @@ import org.springframework.web.bind.annotation.*;
 public class OrderController {
 
     @Autowired
-    private ClientService userService;
+    private ClientService clientService;
+
     @Autowired
     private OrderService orderService;
     @Autowired
@@ -33,44 +35,102 @@ public class OrderController {
 
     // Obtener todas las compras
     @GetMapping("/orders")
-    public List<Order> findOrders() {
+    public List<Order> findOrdersDto() {
         return orderService.findAll();
     }
 
     // Obtener los datos de una compra
     @GetMapping("/orders/{orderId}")
-    public Order findById(@PathVariable long orderId)
-            throws ClientNotFoundException, OrderNotFoundException {
+    public ResponseEntity<Order> findById(@PathVariable long orderId) throws OrderNotFoundException {
         Optional<Order> optionalOrder = orderService.findById(orderId);
-        Order order = optionalOrder.orElseThrow(() -> new OrderNotFoundException(orderId));
-        return order;
+
+        if (optionalOrder.isPresent()) {
+            Order order = optionalOrder.get();
+            return new ResponseEntity<>(order, HttpStatus.OK);
+        } else {
+            throw new OrderNotFoundException(orderId);
+        }
     }
+
+
 
     // Obtener todas las compras de un cliente
     @GetMapping("/client/{clientId}/orders")
-    public List<Order> findClientOrders(@PathVariable long clientId) throws ClientNotFoundException {
-        return orderService.findByClient(clientId);
+    public ResponseEntity<List<Order>> findClientOrders(@PathVariable long clientId) throws ClientNotFoundException {
+        Optional<Client> optionalClient = clientService.findById(clientId);
+
+        if (optionalClient.isPresent()) {
+            List<Order> orders = orderService.findByClient(clientId);
+
+            if (!orders.isEmpty()) {
+                return new ResponseEntity<>(orders, HttpStatus.OK);
+            } else {
+                // El cliente existe pero no tiene compras
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+        } else {
+            // El cliente no existe
+            throw new ClientNotFoundException(clientId);
+        }
     }
+
+
+    // TODO compras de un producto
+    // Obtener todas las compras de un producto
+    //@GetMapping("/product/{productId}/orders")
+    //public List<OrderOutDto> findByProduct(@PathVariable long productId) throws ProductNotFoundException {
+    //    List<Order> ordersWithProduct = orderService.findByProducts(productId);
+
+      //  return listDtos;
+    //}
 
     // Añadir una compra a un cliente
     @PostMapping("/client/{clientId}/orders")
-    public void addOrder(@RequestBody OrderInDto orderInDto, @PathVariable long clientId) throws ClientNotFoundException, ProductNotFoundException {
-        orderService.addOrder(orderInDto, clientId);
+    public ResponseEntity<Void> addOrder(@RequestBody OrderInDto orderInDto, @PathVariable long clientId)
+            throws ClientNotFoundException, ProductNotFoundException {
+        Optional<Client> optionalClient = clientService.findById(clientId);
+
+        if (optionalClient.isPresent()) {
+            orderService.addOrder(orderInDto, clientId);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        } else {
+            // El cliente no existe
+            throw new ClientNotFoundException(clientId);
+        }
     }
 
-    // TODO Hacer el resto de operaciones hasta el CRUD
 
     // Modificar una compra
-    @PutMapping("/order/{ordertId}")
-    public void modifyOrder(@RequestBody Order order, @PathVariable long orderId) {
-        orderService.modifyOrder(order, orderId);
+    @PutMapping("/order/{orderId}")
+    public ResponseEntity<Void> modifyOrder(@RequestBody Order order, @PathVariable long orderId)
+            throws OrderNotFoundException {
+        Optional<Order> optionalOrder = orderService.findById(orderId);
+
+        if (optionalOrder.isPresent()) {
+            orderService.modifyOrder(order, orderId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            // La orden no existe
+            throw new OrderNotFoundException(orderId);
+        }
     }
+
 
     // Eliminar una compra por ID
     @DeleteMapping("/order/{orderId}")
-    public void removeOrder(@PathVariable long orderId) {
-        orderService.removeOrder(orderId);
+    public ResponseEntity<Void> removeOrder(@PathVariable long orderId)
+            throws OrderNotFoundException {
+        Optional<Order> optionalOrder = orderService.findById(orderId);
+
+        if (optionalOrder.isPresent()) {
+            orderService.removeOrder(orderId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            // La orden no existe
+            throw new OrderNotFoundException(orderId);
+        }
     }
+
 
     // Control de excepciones
     @ExceptionHandler(ClientNotFoundException.class)
